requirements.txt
tensorflow==2.0.0
numpy==1.17.5
python==3.7.6
biopython==1.76
scikit-learn==0.22.1

